% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kh3d_basic.R
\name{kh3_breaks}
\alias{kh3_breaks}
\title{Break helper function}
\usage{
kh3_breaks(lim, n = 3)
}
\arguments{
\item{lim}{numeric vector of length 2, range of the axis.}

\item{aspect}{numeric vector of length 3, defining the plot aspect ratio.}
}
\description{
\code{kh3_breaks} is a helper for breaks in kh3d plots.
}
\details{
Helper function to create axis breaks within kh3d.
}
