% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kh3d_basic.R
\docType{data}
\name{kh_clr}
\alias{kh_clr}
\title{kh color vector}
\format{An object of class \code{character} of length 2.}
\usage{
kh_clr
}
\description{
\code{kh_clr} is a combination of colors I like.
}
\details{
A vector containg blue and orange.
}
\keyword{datasets}
