% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kh3d_basic.R
\name{rgl_add_axes}
\alias{rgl_add_axes}
\title{Add 3d axes}
\usage{
rgl_add_axes(x, y, z, axis.col = "grey", xlab = "", ylab = "",
  zlab = "", show.plane = TRUE, show.ticks = TRUE,
  show.bbox = FALSE, bbox.col = c(lpha("#333377", 0.4), "black"))
}
\description{
\code{rgl_add_axes} add 3d axes device.
}
\details{
stolen from \link{http://www.sthda.com/english/wiki/a-complete-guide-to-3d-visualization-device-system-in-r-r-software-and-data-visualization}
}
\examples{

rgl_init(bg = rgb(.3,.3,.3), aspect = aspect)
rgl_add_axes(data$x, data$y, data$z,
            show.bbox = FALSE,
            show.plane = FALSE)

}
