% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kh3d_basic.R
\name{rgl_init}
\alias{rgl_init}
\title{Initialize 3d plot}
\usage{
rgl_init(new.device = FALSE, bg = "white", width = 640,
  aspect = c(1, 1, 1), theta = 0, phi = -80, zoom = 1.03,
  fov = 60)
}
\description{
\code{rgl_init} creates the 3d device.
}
\details{
stolen from \link{http://www.sthda.com/english/wiki/a-complete-guide-to-3d-visualization-device-system-in-r-r-software-and-data-visualization}
}
\examples{

rgl_init()

}
