% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kh3d_basic.R
\name{tib_xyz_to_mat}
\alias{tib_xyz_to_mat}
\title{Convert tibble to rgl matrix}
\usage{
tib_xyz_to_mat(tib)
}
\description{
\code{tib_xyz_to_mat} converts a tible to a matrix.
}
\details{
Data preparation to convert a data tibble
cith three columns (x,y,z) into a matrix for
use with rgl.
}
\examples{

tib_mat <- data \%>\%
  tib_xyz_to_mat()

}
